backend/requirements.txt
django==4.1
djangorestframework==3.13.1
django-cors-headers==3.13.0
keyboard==0.13.5
asyncio==3.4.3
opencv-python==4.6.0.66